% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COD_report.r
\name{COD_Report}
\alias{COD_Report}
\title{COD_Report function}
\usage{
COD_Report(Article_ID, Decision_Errors, Major_Numerical_Errors,
  Minor_Numerical_Errors)
}
\arguments{
\item{Article_ID}{Enter the article's unique ID code}

\item{Decision_Errors}{Enter the number of decision errors}

\item{Major_Numerical_Errors}{Enter the number of major numerical errors}

\item{Minor_Numerical_Errors}{Enter the number of minor numerical errors}
}
\value{
Returns a formatted table (via kable) reporting error tallys. Also writes out a csv object containing the error tallys.
}
\description{
This is a helper function for building standardised final outcoems for COD reproducibility reports.
}
\examples{
COD_Report(Article_ID = "ABhgyo", Decision_Errors = 1, Major_Numerical_Errors = 4, Minor_Numerical_Errors = 12)
}
