% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codReport2.r
\name{codReport2}
\alias{codReport2}
\title{codReport2 function}
\usage{
codReport2(Report_Type, Article_ID, valuesChecked, Total_df = 0,
  Total_p = 0, Total_mean = 0, Total_sd = 0, Total_se = 0,
  Total_ci = 0, Total_bf = 0, Total_t = 0, Total_F = 0, Total_es = 0,
  Total_median = 0, Total_other = 0, Insufficient_Information_Errors,
  Decision_Errors, Major_Numerical_Errors, Minor_Numerical_Errors,
  Major_df = 0, Major_p = 0, Major_mean = 0, Major_sd = 0,
  Major_se = 0, Major_ci = 0, Major_bf = 0, Major_t = 0, Major_F = 0,
  Major_es = 0, Major_median = 0, Major_other = 0,
  affectsConclusion = c("Yes", "No", "Unclear"), errorLocus_typo = 0,
  errorLocus_specification = 0, errorLocus_analysis = 0, errorLocus_data,
  errorLocus_unidentified = 0, Author_Assistance, Nature_of_Assistance = NA,
  correctionSuggested = NA, correctionPublished = NA,
  Resolved_Insufficient_Information_Errors = NA,
  Resolved_Decision_Errors = NA, Resolved_Major_Numerical_Errors = NA)
}
\arguments{
\item{Report_Type}{Enter 'pilot' or 'joint'}

\item{Article_ID}{Enter the article's unique ID code}

\item{Insufficient_Information_Errors}{Enter the number of Insufficient Information Errors}

\item{Decision_Errors}{Enter the number of decision errors}

\item{Major_Numerical_Errors}{Enter the number of major numerical errors}

\item{Minor_Numerical_Errors}{Enter the number of minor numerical errors}

\item{Author_Assistance}{Enter whether author assistance was required (T/F)}
}
\value{
Returns a formatted table (via kable) reporting error tallys. Also writes out a csv object containing the error tallys.
}
\description{
This is a helper function for building standardised final outcomes for COD reproducibility reports.
}
\examples{
codReport2(Report_Type = "final", Article_ID = "ABhgyo", Insufficient_Information_Errors = 0, Decision_Errors = 1, Major_Numerical_Errors = 4, Minor_Numerical_Errors = 12, Author_Assistance = T)
}
