% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareValues.r
\name{compareValues}
\alias{compareValues}
\title{compareValues function}
\usage{
compareValues(reportedValue, obtainedValue, isP = F)
}
\arguments{
\item{reportedValue}{Enter the value reported in the article}

\item{obtainedValue}{Enter the corresponding value obtained in your reproducibility check}
}
\value{
Returns a short text report noting the error type and the PE.
}
\description{
This is a helper function to aid comparison of values reported in articles to values obtained in reproducibility checks.
(1) calculates the percentage error (PE) between a reported value and an obtained value.
(2) identifies the error type (major numerical, minor numerical, no error)
(3) if p-values are being compared, user must set isP to TRUE. Function will identify if there is an additional decision error.
Errors types are defined as follows:
>> 'minor numerical': i.e., >= 2% PE < 10%
>> 'major numerical' (i.e., PE >= 10%)
If p values are being compared, also returns an additional error type:
>> 'decision error' (i.e., reported p and obtained p fall on different sides of the .05 threshold)
}
\examples{
compareValues(reportedValue = 3.45, obtainedValue = 1.34)
compareValues(reportedValue = .054, obtainedValue = .049, isP = T)
compareValues(reportedValue = 15.63, obtainedValue = 15.63)
}
