% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareValues2.r
\name{compareValues2}
\alias{compareValues2}
\title{compareValues2 function}
\usage{
compareValues2(reportedValue, obtainedValue, valueType = c("p", "mean", "sd",
  "se", "df", "F", "t", "bf", "ci", "median", "es", "other"),
  updatedReportObject = reportObject)
}
\arguments{
\item{reportedValue}{Enter the value reported in the article}

\item{obtainedValue}{Enter the corresponding value obtained in your reproducibility check}
}
\value{
Returns a short text report noting the error type and the PE.
}
\description{
This is a helper function to aid comparison of values reported in articles to values obtained in reproducibility checks.
(1) calculates the percentage error (PE) between a reported value and an obtained value.
(2) identifies the error type (decision error, major numerical, minor numerical, no error)
(3) user must specificy the value type from the defaults list (or use 'other' if not listed)
Errors types are defined as follows:
>> 'minor numerical': i.e., >= 2% PE < 10%
>> 'major numerical' (i.e., PE >= 10%)
If p values are being compared, also returns an additional error type:
>> 'decision error' (i.e., reported p and obtained p fall on different sides of the .05 threshold)
}
\examples{
compareValues2(reportedValue = '3.45', obtainedValue = 1.34, valueType = 'mean')
compareValues2(reportedValue = '.054', obtainedValue = .049, valueType = 'p')
compareValues2(reportedValue = '15.63', obtainedValue = 15.63, valueType = 'sd')
}
